lesson_25/project_lesson_25/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-filter==25.2
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.1
psycopg==3.3.2